Adds ability to run multiple minitest tests and compile a report of all executions
All tests are under the test/specs/ folder
Adds a Rakefile in the root of the project. Uses Rake build tool for Ruby projects to define tasks for running tests.
It will run all test files under the test/specs/ folder and generates an HTML report in the html_reports folder